# docs/swagger.yaml
basePath: /
definitions:
...
  hoursBarber.CreateException:
    properties:
      barbeiro_id:
        type: integer
      data_execeção:
        type: string
      motivo:
        type: string
    required:
    - barbeiro_id
    - data_execeção
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: This is a sample server Petstore server.
  title: Barber Shop API
  version: "1.0"
paths:
...
  /barber/hoursBarberException/create:
    post:
      consumes:
      - application/json
      description: Cria uma exceção de hora de trabalho para o barbeiro
      parameters:
      - description: Create hours barber exception
        in: body
        name: hoursBarberException
        required: true
        schema:
          $ref: '#/definitions/hoursBarber.CreateException'
      produces:
      - application/json
      responses:
        "200":
          description: Sem conteúdo
      summary: Criação de exceções de hora de trabalho para o barbearia
      tags:
      - hoursBarber
  /barber/hoursBarberException/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Remove uma exceção de hora de trabalho para o barbeiro
      parameters:
      - description: Exception ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Barber hours exception deleted successfully
      summary: Remoção de exceção de hora de trabalho
      tags:
      - hoursBarber
...
swagger: "2.0"